requirements.txt
Flask
flask-cors
pymongo
pillow
python-dotenv
waitress
flask-mailc